requirements.txt
absl-py==1.3.0
aiofiles==22.1.0
apturl==0.5.2
astunparse==1.6.3
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.2.0
certifi==2022.9.24
chardet==4.0.0
charset-normalizer==2.1.1
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
fasteners==0.14.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==2.17.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
gyp==0.1
h11==0.9.0
h5py==3.7.0
html5lib==1.1
html5tagger==1.3.0
httpcore==0.11.1
httplib2==0.20.2
httptools==0.5.0
httpx==0.15.4
idna==3.4
importlib-metadata==4.6.4
jax==0.4.8
jeepney==0.7.1
jsons==1.6.3
kazam==1.4.5
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==14.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.1
ml-dtypes==0.0.4
monotonic==1.6
more-itertools==8.10.0
multidict==5.0.0
netifaces==0.11.0
numpy==1.23.3
oauthlib==3.2.1
olefile==0.46
opencv-contrib-python==4.6.0.66
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.2.0
protobuf==3.19.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyGObject-stubs==1.0.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-version==0.0.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
rsa==4.9
sanic==20.12.7
Sanic-Cors==0.10.0
sanic-ext==23.3.0
sanic-jwt==1.8.0
Sanic-Plugins-Framework==0.9.5
sanic-routing==0.7.2
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
secure==0.3.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
systemd-python==234
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.3
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tracerite==1.1.0
typing_extensions==4.4.0
typish==1.9.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.5.0
unattended-upgrades==0.1
urllib3==1.26.12
usb-creator==0.3.7
uvloop==0.17.0
wadllib==1.3.6
webencodings==0.5.1
websockets==10.1
Werkzeug==2.2.2
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
